+++ source/libtomo/recon/bart.c
@@ -48,7 +48,7 @@
 void
 bart(const float* data, int dy, int dt, int dx, const float* center, const float* theta,
      float* recon, int ngridx, int ngridy, int num_iter, int num_block,
-     const float* ind_block)  // TODO: I think ind_block should be int*
+     const int* ind_block)
 {
     if(dy == 0 || dt == 0 || dx == 0)
         return;
